+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - GP2-341 - Adaptation language info and stats update
 - GP2-1270 - income data for economy tab
+- GP2-1318 - Compare markets - refactor and loading twinkle
 - GP2-1265 - Compare markets - society data attribution
 - GP2-1258 - Compare markets - integrate society data
 - GP2-1270 - income data for economy tab